refactor: 공통 확인 다이얼로그 DialogUtils로 분리 #31
